requirements.txt
manim
google-generativeai
python-dotenv
